+++ package.json
@@ -4,7 +4,7 @@
   "version": "0.0.1",
   "author": "",
   "dependencies": {
-    "@fastify/static`": "6.12.0",
+    "@fastify/static": "6.12.0",
     "@langchain/core": "0.1.24",
     "@langchain/openai": "0.0.14",
     "@nestjs/common": "^10.0.0",
@@ -13,7 +13,7 @@
     "@nestjs/platform-fastify": "10.3.2",
     "@nestjs/swagger": "7.2.0",
     "@nestjs/typeorm": "10.0.1",
-    "@tsconfig/node18": "18.2.  2",
+    "@tsconfig/node18": "18.2.2",
     "class-transformer": "0.5.1",
     "class-validator": "0.14.1",
     "http-status-codes": "2.3.0",
