+++ src/e2e/articles/get-all.e2e-spec.ts
@@ -1,5 +1,5 @@
 import { getTestUser } from 'e2e/lib/get-test-user';
-import { RequestClient } from 'lib/request-factory';
+import { RequestClient } from 'lib/request';
 
 describe('GET v1/memories/articles', () => {
   let user: { uuid: string; apiKey: string };
