+++ src/lib/validate.ts
@@ -9,7 +9,6 @@ export const validate = <T, E extends Error>(
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   const validationErrors = validateSync(entity as any);
   if (validationErrors.length > 0) {
-    console.log(JSON.stringify(validationErrors, null, 2));
     throw (
       // eslint-disable-next-line @typescript-eslint/no-throw-literal
       err?.(validationErrors) ??
